+++ modules/tinymce/src/core/main/ts/api/Options.ts
@@ -573,7 +573,8 @@ const register = (editor: Editor) => {
   });
 
   registerOption('element_format', {
-    processor: 'string'
+    processor: 'string',
+    default: 'html'
   });
 
   registerOption('entities', {