+++ modules/tinymce/src/plugins/table/test/ts/browser/ClipboardTest.ts
@@ -582,7 +582,7 @@ describe('browser.tinymce.plugins.table.ClipboardTest', () => {
       cleanTableHtml(editor.getContent()),
 
       '<table>' +
-      '<colgroup><col /><col /><col /></colgroup>' +
+      '<colgroup><col><col><col></colgroup>' +
       '<tbody>' +
       '<tr><td>2</td><td>1</td><td>2</td></tr>' +
       '<tr><td>3</td><td>2</td><td>3</td></tr>' +
@@ -612,7 +612,7 @@ describe('browser.tinymce.plugins.table.ClipboardTest', () => {
       cleanTableHtml(editor.getContent()),
 
       '<table>' +
-      '<colgroup><col /><col /><col /><col /></colgroup>' +
+      '<colgroup><col><col><col><col></colgroup>' +
       '<tbody>' +
       '<tr><td>1</td><td>2</td><td>1</td><td>3</td></tr>' +
       '<tr><td>2</td><td>3</td><td>2</td><td>4</td></tr>' +
@@ -643,7 +643,7 @@ describe('browser.tinymce.plugins.table.ClipboardTest', () => {
       cleanTableHtml(editor.getContent()),
 
       '<table data-snooker-locked-cols="0">' +
-      '<colgroup><col /><col /><col /></colgroup>' +
+      '<colgroup><col><col><col></colgroup>' +
       '<tbody>' +
       '<tr><td>1</td><td>2</td><td>2</td></tr>' +
       '<tr><td>2</td><td>3</td><td>3</td></tr>' +
@@ -674,7 +674,7 @@ describe('browser.tinymce.plugins.table.ClipboardTest', () => {
       cleanTableHtml(editor.getContent()),
 
       '<table data-snooker-locked-cols="2">' +
-      '<colgroup><col /><col /><col /></colgroup>' +
+      '<colgroup><col><col><col></colgroup>' +
       '<tbody>' +
       '<tr><td>1</td><td>1</td><td>2</td></tr>' +
       '<tr><td>2</td><td>2</td><td>3</td></tr>' +